+++ domain/ingester.go
@@ -2,7 +2,7 @@ package domain
 
 type GRPCIngesterConfig struct {
 	// Flag to enable the GRPC ingester server
-	Enabeld bool `mapstructure:"enabled"`
+	Enabled bool `mapstructure:"enabled"`
 
 	// The maximum number of bytes to receive in a single GRPC message
 	MaxReceiveMsgSizeBytes int `mapstructure:"max-receive-msg-size-bytes"`
@@ -12,4 +12,14 @@ type GRPCIngesterConfig struct {
 
 	// The number of seconds to wait for a connection to the server.
 	ServerConnectionTimeoutSeconds int `mapstructure:"server-connection-timeout-seconds"`
+}
+
+// BlockPoolMetadata contains the metadata about unique pools
+// and denoms modified in a block.
+type BlockPoolMetadata struct {
+	// DenomLiquidityMap is a map of denoms to their liquidities across pools.
+	// These are constructed from the pool IDs updated within a block.
+	DenomLiquidityMap DenomLiquidityMap
+	// PoolIDs are the IDs of all pools updated within a block.
+	PoolIDs map[uint64]struct{}
 }