+++ router/usecase/optimized_routes_test.go
@@ -178,7 +178,7 @@ func (s *RouterTestSuite) TestGetBestSplitRoutesQuote() {
 				}),
 			},
 
-			maxSplitIterations: 17,
+			maxSplitIterations: 10,
 
 			tokenIn: sdk.NewCoin(DenomTwo, sdk.NewInt(56_789_321)),
 
@@ -658,7 +658,7 @@ func (s *RouterTestSuite) TestGetCustomQuote_GetCustomDirectQuote_Mainnet_UOSMOU
 	routerRepositoryMock.SetTakerFees(mainnetState.TakerFeeMap)
 
 	// Setup pools usecase mock.
-	poolsUsecase := poolsusecase.NewPoolsUsecase(&domain.PoolsConfig{}, "node-uri-placeholder", routerRepositoryMock)
+	poolsUsecase := poolsusecase.NewPoolsUsecase(&domain.PoolsConfig{}, "node-uri-placeholder", routerRepositoryMock, domain.UnsetScalingFactorGetterCb)
 	poolsUsecase.StorePools(mainnetState.Pools)
 
 	routerUsecase := routerusecase.NewRouterUsecase(routerRepositoryMock, poolsUsecase, config, emptyCosmWasmPoolsRouterConfig, &log.NoOpLogger{}, cache.New(), cache.New())