+++ tests/quote_response.py
@@ -1,0 +1,41 @@
+from decimal import *
+
+# Coin represents a coin in the /router/quote response
+class Coin:
+    def __init__(self, denom, amount):
+        self.denom = denom
+        self.amount = int(amount)
+
+# Pool represents a pool in the /router/quote response
+class Pool:
+    def __init__(self, id, type, balances, spread_factor, token_out_denom, taker_fee, **kwargs):
+        self.id = int(id)
+        self.type = type
+        self.balances = balances
+        self.spread_factor = float(spread_factor)
+        self.token_out_denom = token_out_denom
+        self.taker_fee = float(taker_fee)
+        code_id = kwargs.get('code_id', 0)
+        # Only CW pools have code id
+        if code_id:
+            self.code_id = int(code_id)
+
+# Route represents a route in the /router/quote response
+class Route:
+    def __init__(self, pools, out_amount, in_amount, **kwargs):
+        self.pools = [Pool(**pool) for pool in pools]
+        self.out_amount = int(out_amount)
+        self.in_amount = int(in_amount)
+        # "has-cw-pool" format is unsupported
+        self.has_cw_pool = kwargs.get('has-cw-pool', False)
+
+# QuoteResponse represents the response format
+# of the /router/quote endpoint
+class QuoteResponse:
+    def __init__(self, amount_in, amount_out, route, effective_fee, price_impact, in_base_out_quote_spot_price):
+        self.amount_in = Coin(**amount_in)
+        self.amount_out = int(amount_out)
+        self.route = [Route(**r) for r in route]
+        self.effective_fee = Decimal(effective_fee)
+        self.price_impact = Decimal(price_impact)
+        self.in_base_out_quote_spot_price = Decimal(in_base_out_quote_spot_price)